requirements.txt
numpy==1.25.2
pandas==2.0.3
# Pin scikit‑learn to a specific version to ensure reproducibility
scikit-learn==1.3.2
xgboost==2.0.0
matplotlib==3.7.2
pyyaml==6.0
joblib==1.3.0
imbalanced-learn==0.11.2
seaborn==0.13.0

# Additional optional dependencies used in the extended pipeline:
#   - flaml for AutoML search (automl.py)
#   - lifelines for survival modelling (advanced_models.py)
#   - scipy for ODE simulations (ode_simulation.py)
#   - umap-learn for manifold learning in eda.py
#   - shap for SHAP value interpretability (interpret_shap.py)
#   - goatools for GO enrichment analysis
#   - pytest for unit tests and CI
flaml==2.0.0
lifelines==0.30.0
scipy==1.11.0
umap-learn==0.5.5
shap==0.43.0
goatools==1.3.1
pytest==7.4.0

# New optional dependencies for advanced features and reproducibility
gprofiler-official==1.0.0
GEOparse==2.0.0
networkx==3.1
optuna==3.5.0
torch-geometric>=2.3.0
torch==2.0.1
transformers==4.32.0
pytorch-lightning==2.1.0
papermill==2.4.0
plotly==5.15.0

# Additional dependencies for survival analysis, causal inference, fairness auditing
scikit-survival==0.22.0
dowhy==0.9.2
aif360==0.5.0
combat-py==0.0.2
scikit-multilearn==0.2.0